requirements.txt
aiofiles==0.6.0
aiohttp==3.6.2
asgiref==3.2.10
bcrypt==3.2.0
confluent-kafka==1.9.2
cx_Oracle==8.3.0
fastapi==0.62.0
numpy==1.21.1
opencv-python==4.5.3.56
pandas==1.1.2
pdf2image==1.14.0
Pillow==7.1.2
pydantic==1.7.3
PyJWT==1.7.1
requests==2.23.0
SQLAlchemy==1.4.28
starlette==0.13.6
typing-extensions==3.7.4.3
pycryptodome==3.15.0
python-multipart==0.0.5
python-socketio==5.7.2
grpcio==1.53.0
grpcio-tools==1.53.0